+++ src/constants/inputs.ts
@@ -1,13 +1,11 @@
+export const DEFAULT_IGNORE_GLOB = "";
+export const DEFAULT_IS_DRY_RUN = false;
+export const DEFAULT_SVGO_V1_CONFIG_PATH = ".svgo.yml";
+export const DEFAULT_SVGO_V2_CONFIG_PATH = "svgo.config.js";
+export const DEFAULT_SVGO_VERSION = 2;
+
 export const INPUT_NAME_DRY_RUN = "dry-run";
 export const INPUT_NAME_IGNORE = "ignore";
 export const INPUT_NAME_REPO_TOKEN = "repo-token";
 export const INPUT_NAME_SVGO_CONFIG = "svgo-config";
 export const INPUT_NAME_SVGO_VERSION = "svgo-version";
-
-export const INPUT_NAMES = {
-  dryRun: INPUT_NAME_DRY_RUN,
-  ignore: INPUT_NAME_IGNORE,
-  repoToken: INPUT_NAME_REPO_TOKEN,
-  svgoConfig: INPUT_NAME_SVGO_CONFIG,
-  svgoVersion: INPUT_NAME_SVGO_VERSION,
-};